Migrate from SDL 2.0.14 to SDL 3.4.14
Repoint the ext/sdl submodule at release-3.4.14 and port the SDL usage in
MppAppSupport, DemoSuite and PipelineEditor to the SDL3 API.

Build:
- SDL_TEST becomes SDL_TEST_LIBRARY; also disable SDL_TESTS/SDL_EXAMPLES/SDL_INSTALL.
- Link SDL3::SDL3 instead of SDL2, and drop MPP_SDL_INCLUDE_DIR: headers are now
  included as <SDL3/SDL.h> and the include path comes from the target.
- Drop SDL2main. Both apps define WinMain themselves, so it was never used, and
  SDL3 has no equivalent library target.
- Runtime copies now use the SDL3-shared target.

API:
- Event constants renamed (SDL_EVENT_*); window events are top-level rather than
  nested under SDL_WINDOWEVENT.
- Key events flattened: evt.key.keysym.sym/scancode/mod -> evt.key.key/scancode/mod.
- Letter keycodes uppercased (SDLK_a -> SDLK_A), SDLK_BACKQUOTE -> SDLK_GRAVE,
  KMOD_* -> SDL_KMOD_*.
- System cursor enums renamed; SDL_FreeCursor -> SDL_DestroyCursor;
  SDL_ShowCursor(bool) -> SDL_ShowCursor()/SDL_HideCursor().
- SDL_Init and SDL_GL_MakeCurrent return bool instead of an int status.
- SDL_GL_DeleteContext -> SDL_GL_DestroyContext.
- SDL_GL_GetDrawableSize -> SDL_GetWindowSizeInPixels.
- SDL_SetTextInputRect -> SDL_SetTextInputArea, which is window-scoped.
- Mouse position and warp coordinates are floats.
- Display enumeration reworked onto SDL_GetDisplays and
  SDL_GetFullscreenDisplayModes.
- SDL_syswm.h is gone; the file dialog gets its HWND from the window properties.

Behaviour preserved or corrected:
- SDL3 leaves text input disabled by default, so create() now calls
  SDL_StartTextInput.
- An SDL3 fullscreen window is borderless-desktop unless given an explicit mode,
  so create() pins the closest mode to the requested size.
- The vsync argument previously set SDL_RENDERER_PRESENTVSYNC on a renderer flag
  variable that was never used, so it had no effect. Those SDL2 constants are gone
  in SDL3; vsync is now applied via SDL_GL_SetSwapInterval, which this
  GL-presented window actually honours.

+++ mpp-app-support/src/TimerSDL.cpp
@@ -1,4 +1,4 @@
-#include <SDL.h>
+#include <SDL3/SDL.h>
 #include "mpp/app/TimerSDL.h"
 
 TimerSDL::TimerSDL() :
@@ -9,13 +9,14 @@ TimerSDL::TimerSDL() :
 
 void TimerSDL::reset()
 {
-	mStartTime = SDL_GetTicks() / 1000.0f;
+	// SDL3 returns 64-bit milliseconds.
+	mStartTime = (float)(SDL_GetTicks() / 1000.0);
 	mRunningTime = 0.0f;
 }
 
 float TimerSDL::getDeltaTime() const
 {
-	float newTime = SDL_GetTicks() / 1000.0f;
+	float newTime = (float)(SDL_GetTicks() / 1000.0);
 	float frameTime = newTime - mStartTime;
 
 	mStartTime = newTime;
